Olympic data set analysis
For my analysis, I have transformed, cleaned, analyzed data all with the hope of looking into the subject of if the Olympic games is a gender-balanced event. I have created various charts to show which countries push for women's participation in the Olympic games, and what sports would see more improvement when it comes to women's participation. With this data provided from Kaggle, I have been able to determine that the Olympic games is gender-balanced or at least are becoming a gender-balanced event.  
All activities in this project were conducted with R and R Studio.